requirements.txt
pyrogram
yt-dlp
python-dotenv
flask
tgcrypto